All the changes have been taken in account and i have also arrange some stuff to make them better.
