+++ src/app/(app)/(Policyholder)/assures/[id]/page.tsx
