+++ src/app/api/policyholder/route.ts
@@ -1,17 +1,15 @@
 import { NextRequest, NextResponse } from 'next/server';
-import { getPolicyholders } from '../../../../services/policyholder/read';
+import { getPolicyholders } from '../../../services/policyholder';
 import { currentUser } from "@/lib/auth";
 
 export async function GET(request: NextRequest) {
     const user = await currentUser()
 
     if (!user) {
-        return new Response("You must be logged in", {
-            status: 401,
-            headers: {
-                "content-type": "application/json",
-            },
-        });
+        return NextResponse.json(
+            { error: "You must be logged in" },
+            { status: 401 }
+        );
     }
 
     const searchParams = request.nextUrl.searchParams;
